feat: providedIn Root
You can now share providers over different modules. If you want to have the same instance, even if the provider is defined in two different modules, you can use providedIn in the Injectable decorator and set it to 'root'. If the scope is set to singleton the instance will not be created in the module context but in the root context (which is new!)