refactor: Removed the pipelines API for a more dynamic API
You can now call the update method on the World interface to update your world. There is an optional scope Parameter which you can use to only update a part of your world. See the documentation for more details

BREAKING CHANGE: The init method of a module does no longer have the Pipeline parameter. In fact, this API has been completely removed and is no longer supported, as it has been replaced by a more dynamic approach on which I will work from now on.